Housekeeping and small error fixes
I noticed that the only thing holding me back from running this with certain versions of the modules used here was some argument related issues! I did my best to track all of them down here.

I also appended *.pt files to the .gitignore to ensure not models are accidentally uploaded

+++ toolbox/ui.py
@@ -381,7 +381,7 @@ class UI(QDialog):
         self.app.processEvents()
 
     def set_loading(self, value, maximum=1):
-        self.loading_bar.setValue(value * 100)
+        self.loading_bar.setValue(int(value * 100))
         self.loading_bar.setMaximum(maximum * 100)
         self.loading_bar.setTextVisible(value != 0)
         self.app.processEvents()
